server/requirements.txt
flask
flask-cors
python-dotenv
requests
chromadb
pdfplumber
python-docx
python-pptx
pandas
openpyxl
openai
sentence-transformers
werkzeug
